Added factory activated ProxyMiddleware
